+++ src/ops/splice.rs
@@ -1,7 +1,8 @@
-use crate::any_vec_ptr::IAnyVecRawPtr;
+use crate::any_vec_ptr::{AnyVecPtr, AnyVecRawPtr, IAnyVecRawPtr};
 use crate::{any_vec_ptr, Iter};
 use crate::any_value::AnyValue;
 use crate::ops::iter::Iterable;
+use crate::traits::Trait;
 
 pub struct Splice<'a, AnyVecPtr: IAnyVecRawPtr, ReplaceIter: ExactSizeIterator>
 where
@@ -117,3 +118,36 @@ where
         }
     }
 }
+
+#[allow(suspicious_auto_trait_impls)]
+unsafe impl<'a, Traits: ?Sized + Send + Trait, ReplaceIter: ExactSizeIterator> Send
+for
+    Splice<'a, AnyVecPtr<Traits>, ReplaceIter>
+where
+    ReplaceIter::Item: AnyValue + Send
+{}
+
+#[allow(suspicious_auto_trait_impls)]
+unsafe impl<'a, Type: Send, ReplaceIter: ExactSizeIterator> Send
+for
+    Splice<'a, AnyVecRawPtr<Type>, ReplaceIter>
+where
+    ReplaceIter::Item: AnyValue + Send
+{}
+
+
+#[allow(suspicious_auto_trait_impls)]
+unsafe impl<'a, Traits: ?Sized + Sync + Trait, ReplaceIter: ExactSizeIterator> Sync
+for
+    Splice<'a, AnyVecPtr<Traits>, ReplaceIter>
+where
+    ReplaceIter::Item: AnyValue + Sync
+{}
+
+#[allow(suspicious_auto_trait_impls)]
+unsafe impl<'a, Type: Sync, ReplaceIter: ExactSizeIterator> Sync
+for
+    Splice<'a, AnyVecRawPtr<Type>, ReplaceIter>
+where
+    ReplaceIter::Item: AnyValue + Sync
+{}